Now catching multiple errors in crawl.sh.

+++ misc/crawl.sh
@@ -7,7 +7,7 @@ echo
 echo
 echo "By Lee Baird"
 echo
-echo "Returns a list of IP external web servers that are linked from home page."
+echo "Find all of the subdomains linked to the homepage."
 echo
 echo "Usage: target.com"
 echo
@@ -16,7 +16,7 @@ read -p "Domain: " domain
 
 if [ -z $domain ]; then
      echo
-     echo "#########################"
+     echo "========================================"
      echo
      echo "Invalid choice."
      echo
@@ -24,18 +24,25 @@ if [ -z $domain ]; then
 fi
 
 echo
-echo "#########################"
+echo "========================================"
 echo
 
 wget -q www.$domain
 
-grep 'href=' index.html | cut -d '/' -f3 | grep $domain | cut -d '"' -f1 | sort -u > tmp
+if [ ! -e index.html ]; then
+     echo
+     exit
+fi
+
+grep 'href=' index.html | cut -d '/' -f3 | grep $domain | cut -d '"' -f1 > tmp
 
 for x in $(cat tmp); do
      host $x | grep 'has address' | cut -d ' ' -f1,4 >> tmp2
 done
 
-column -t tmp2 | sort -u
+if [ -e tmp2 ]; then
+     cat tmp2 | grep $domain | column -t | sort -u
+fi
 
 rm index.html tmp*
